Add support for derived compilers to CompilerId
When a compiler is based on another "parent" compiler and has
very similar functionality, the parent compiler field in
CompilerId can be used to make both the impl(compiler) and
the impl(parent) flags evaluate to true, each for their respective
version.

This can make conditions in packages significantly simpler.